Add workspaces to OpenAPI spec
