+++ src/codegate/dashboard/dashboard.py
@@ -1,5 +1,4 @@
 import asyncio
-import json
 from typing import AsyncGenerator, List, Optional
 
 import structlog
@@ -60,18 +59,3 @@ async def stream_sse():
     Send alerts event
     """
     return StreamingResponse(generate_sse_events(), media_type="text/event-stream")
-
-
-def generate_openapi():
-    # Create a temporary FastAPI app instance
-    app = FastAPI()
-
-    # Include your defined router
-    app.include_router(dashboard_router)
-
-    # Generate OpenAPI JSON
-    openapi_schema = app.openapi()
-
-    # Convert the schema to JSON string for easier handling or storage
-    openapi_json = json.dumps(openapi_schema, indent=2)
-    print(openapi_json)
